Internal: force-evaluate the new render state.
Widgets might retain computations which are not actually needed to render
the image (e. g., the text widget might contain more lines than can be shown
on screen). This seems to leak space as it is left unclear that computations
needed to render the current frame will not be needed anymore after the
image is shown.