str: avoid thunk-buildups.
When processing lines, only reference computations which are needed
in the given context. In particular, don’t build up thunks for columns
and lines which cannot be shown.

While «renderFinal» forces unused computations to be garbage-collected,
as of 8e1d7f4, not building up thunks here will save the GC some work.