Fix selection handling in listReplace
The previous implementation of the merge function was buggy and could
for example return out of bounds indices even for non-empty lists
and generally didn't behave as intended.

This new implementation should be better. It now returns -1 if the
replacement list is empty, but this case is excluded on the caller side
in listReplace.