Fix bug in adjusting viewport with 'visible'.
If the size of the visibility request was larger than the available
space, then the rendering of a viewport was toggling between two states,
one with aligning on the end of the visibility request, and another one
aligning on the start. This commit fixes it so that a visibility request
is always aligned on the start if not enough space is available.