List: abstract container type
Introduce 'GenericList' which abstracts list behaviour over the
container type.  This allows the use of types other than Vector, if
it would suit a particular application (e.g. lazy loading of list
contents).

'List' is retained as a type synonym for backwards compatibility.
For a container type to be usable with 'List', it must have
instances of 'Traversable' and a new type class 'Splittable' which
allows splitting the container at a given index.

Some operations impose additional constraints on the container type:

- listInsert: 'Applicative' and 'Semigroup'
- listRemove: 'Semigroup'
- listClear: 'Monoid'

Fixes: https://github.com/jtdaugherty/brick/issues/201