List: add "move to next" function
A common use case is to seek to the next occurrence in the list of
an element matching some predicate (e.g. next unread email).  Add
the 'listFindBy' function to satisfy this use case.

'listFindBy' is lazy.  Also redefine 'listMoveToElement' in terms of
'listFindBy', making it lazy.  This changes the constraints from
(Traversable t) to (Foldable t, Splittable t), which is reasonable.