AttrMap: merge styles while folding attribute map lookup results
This change fixes a bug where a more general attribute map entry applied
a style that was not inherited by more specific map entries. This bug
led to the more specific entry's style taking precedence over the style
of the less specific entry, which is not the expected behavior. This
change fixes that behavior so that less specific entry styles are merged
with more specific entry styles.