takeColumns: take even zero-width characters as long as possible
This change makes it so that takeColumns (which is used by 'str' to
account for the widths of Unicode characters) will continue taking
characters that do not exhaust the allowed width, meaning that it will
now take zero-width characters. Previously it stopped taking characters
as soon as it exhausted the allowed width and that meant it would ignore
zero-width characters following that point.