customMain: restore initial terminal input state on shutdown
This change ensures that the initial terminal input state flags are
restored on shutdown (regardless of exceptions) even if changes to the
state flags have occurred via 'suspendAndResume'. Prior to this change,
breaking changes to the input state flags in the last 'suspendAndResume'
before program exit were propagated to the end user's terminal
environment which could lead to broken or garbled terminal I/O.

For more details, see

https://github.com/jtdaugherty/vty/issues/187

This change also increases the lower bound on 'vty' to get access to the
API required to support this behavior.