Extent: remove unused extentOffset field
This change removes the unused extentOffset field of Extent, which was
being used in some calculations but was never changing from its
initialized value of zero (and thus had no meaningful impact on any of
its uses). I'm not sure what I intended with this field when it was
originally added, but clearly this is dead code.