Include clickable elements in the rendering cache.
Mouse events for `Widget`s were being received inconsistently if/when
they are cached. This appears to be because `Widget`s when rendered
update the list of clickable elements that `brick` should track for
mouse activity. However, when a `Widget` is cached, this information is
lost.

This commit updates the rendering cache to also include the clickable
elements exposed by a `Widget` and update `brick`'s tracking list
appropriately when utilizing the cache.