Brick.Main.runWithVty (internal): fix bug where rendering state was reset by mistake
This change fixes a bug where clickable regions in the UI could be
forgotten by the rendering loop if the application called
"continueWithoutRedraw." This was because although the rendering loop
correctly skipped the redraw, it would reset the rendering state's
clickable regions. That meant that if a mouse click event came in
immediately after a call to "continueWithoutRedraw", the mouse click
would not match any extents because the resource name associated with
the click was not considered clickable. This change fixes runWithVty so
that the rendering state's clickable name list is NOT reset in the
specific case where we're going to skip a screen redraw due to a call to
continueWithoutRedraw.