viewport: fix non-scroll direction width/height in the presence of scroll bars
Prior to this change, the size of a viewport's contents would be off by
one row or column if a scroll bar was enabled because we weren't
accounting for the extra row or column taken up by the bar. This change
makes it so that viewports that scroll only in one direction apply a
width or height constraint to their contents to account for the reduced
rendering area when a scroll bar is enabled.