Table: fill and pad empty cells to avoid layout issues (fixes #369)
This change makes it so that when cells are empty (have zero width and
height) and are top- and left-aligned, we pad them to take up the same
amount of width and height as their row- and column-wise neighbors. This
was already the case for other alignment settings, but the top- and
left-aligned cases were not getting padded.