renderBox: avoid O(n²) slowdowns
`renderBox` computes translations by summing the primary dimensions
of widgets preceding the current widget.  This has O(n²) complexity
and causes poor performance with long lists of widgets.  In one
purebred scenario `renderBox` contributes ~30% of the CPU time of
the program.

Use a State computation to accumulate the total offset during the
traversal.  In the scenario described above, this change caused
`renderBox` to drop to ~4% of the program CPU time.

Also remove an occurence of `sum` that calculateed the remaining
primary dimension space after rendering of greedy widgets.  We
already calculate this value during the greedy widget render
traversal, so reuse it.