Brick.Widgets.Internal: rewrite cropExtents to fix bug for extents with negative upper-left corner coordinates
Prior to this change, cropExtents would fail to crop (and even drop)
extents when they had a starting row or column that was outside of the
context. This led to cropResultToContext not properly removing extents
from a Result when it should have, which led to application bugs where
extents would stick around despite the fact that rendering should have
removed them. This would lead to, for example, regions of a UI still
responding to mouse clicks even though those regions were showing things
that either shouldn't report extents or should report *different*
extents.